Fix comparison of integers of different signs

+++ runtime/platform/utils.h
@@ -364,7 +364,7 @@ class Utils {
 
   template <typename T>
   DART_FORCE_INLINE static bool TestBit(T mask, intptr_t position) {
-    ASSERT(position < sizeof(T) * kBitsPerByte);
+    ASSERT(position < static_cast<intptr_t>(sizeof(T) * kBitsPerByte));
     return ((mask >> position) & 1) != 0;
   }
